+++ io/src/main/java/com/itextpdf/io/font/otf/ActualTextIterator.java
@@ -123,7 +123,7 @@ public class ActualTextIterator implements Iterator<GlyphLine.GlyphLinePart> {
                 break;
             }
             // TODO zero glyph is a special case. Unicode might be special
-            toUnicodeMapResult.append(TextUtil.convertFromUtf32((int) currentGlyph.getUnicode()));
+            toUnicodeMapResult.append(TextUtil.convertFromUtf32(currentGlyph.getUnicode()));
         }
 
         return needsActualText || !toUnicodeMapResult.toString().equals(glyphLinePart.actualText);
