+++ io/src/main/java/com/itextpdf/io/font/otf/GlyphLine.java
@@ -107,7 +107,7 @@ public class GlyphLine {
                     if (glyphs.get(i).getChars() != null) {
                         str.append(glyphs.get(i).getChars());
                     } else if (glyphs.get(i).hasValidUnicode()) {
-                        str.append(TextUtil.convertFromUtf32((int) glyphs.get(i).getUnicode()));
+                        str.append(TextUtil.convertFromUtf32(glyphs.get(i).getUnicode()));
                     }
                 }
             }
@@ -183,7 +183,7 @@ public class GlyphLine {
         if (currentGlyph.getChars() != null) {
             chars.append(currentGlyph.getChars());
         } else if (currentGlyph.hasValidUnicode()) {
-            chars.append(TextUtil.convertFromUtf32((int) currentGlyph.getUnicode()));
+            chars.append(TextUtil.convertFromUtf32(currentGlyph.getUnicode()));
         }
 
         for (int j = 0; j < rightPartLen; ++j) {
@@ -192,7 +192,7 @@ public class GlyphLine {
             if (currentGlyph.getChars() != null) {
                 chars.append(currentGlyph.getChars());
             } else if (currentGlyph.hasValidUnicode()) {
-                chars.append(TextUtil.convertFromUtf32((int) currentGlyph.getUnicode()));
+                chars.append(TextUtil.convertFromUtf32(currentGlyph.getUnicode()));
             }
             removeGlyph(gidx.idx--);
         }
@@ -210,9 +210,9 @@ public class GlyphLine {
         if (oldGlyph.getChars() != null) {
             newGlyph.setChars(oldGlyph.getChars());
         } else if (newGlyph.hasValidUnicode()) {
-            newGlyph.setChars(TextUtil.convertFromUtf32((int) newGlyph.getUnicode()));
+            newGlyph.setChars(TextUtil.convertFromUtf32(newGlyph.getUnicode()));
         } else if (oldGlyph.hasValidUnicode()) {
-            newGlyph.setChars(TextUtil.convertFromUtf32((int) oldGlyph.getUnicode()));
+            newGlyph.setChars(TextUtil.convertFromUtf32(oldGlyph.getUnicode()));
         }
         glyphs.set(idx, newGlyph);
     }
